+++ scalability_explorer/src/main/java/eu/choreos/vv/increasefunctions/LinearIncrease.java
@@ -5,7 +5,7 @@ public class LinearIncrease implements ScalabilityFunction {
 
 	@Override
 	public Number increaseParams(Number actualParam, Number fixedParam) {
-		return actualParam.intValue() + fixedParam.intValue();
+		return actualParam.doubleValue() + fixedParam.doubleValue();
 	}
 
 }
